scripts/requirements-cluster.txt
# 离线聚类：HF 服装数据 → UMAP → HDBSCAN → lib/cluster-data.json
datasets>=2.14.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
requests>=2.28.0
umap-learn>=0.5.5
hdbscan>=0.8.33
scikit-learn>=1.3.0
